create data obj in sampleDataModels.js with sampleDataMethods
