--- database/sampleData.js
@@ -1,3 +1,0 @@
-var { getRandomInt, type, pageNum, publisher, dates, isbn, language, characterName, award, coverUrl} = require('./sampleDataModels.js');
-
-console.log(getRandomInt(0, 10));
